# packages/contracts/openapi/collections-v1.yaml
openapi: 3.1.0
info:
  title: DeepDocParse Explicit Collection Catalog
  version: 1.0.0
  description: See ../ddp/collection-catalog-format.md. Corpus owns collections and all source authorization.
security: [{session: []}, {userKey: []}]
paths:
...
  /internal/federation/published-collections:
    get:
      operationId: enumerateNodePublishedCollections
      security: [{service: []}]
      description: |
        Service-only, **node-scoped** producer for peer directory reads. The caller supplies
        no scope and no caller identity: the fixed peer scope, caller digest and origin are
        derived inside corpus from this node's persistent identity. Empty node identity fails
        closed. The caller-scoped /internal/federation/collections path above is unchanged.
      parameters:
        - {name: snapshot_id, in: query, schema: {type: string, maxLength: 32}}
        - {name: cursor, in: query, schema: {type: string, maxLength: 32}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
      responses:
        '200': {description: Stable node catalog page with separate empty terminal proof, content: {application/json: {schema: {$ref: '#/components/schemas/CatalogPage'}}}}
        '401': {description: Missing service credentials or actor context}
        '403': {description: Non-service actor}
        '409': {description: catalog_snapshot_changed; never a revocation}
        '410': {description: catalog_snapshot_expired or catalog_snapshot_invalid with the original revoked identities}
        '429': {description: Node snapshot storage bound reached}
        '503': {description: Persistent node identity unavailable; never emit origin-less descriptors}
        '507': {description: Authorized catalog too large; never claim complete}
  /api/v1/federation/collections:
    get:
      operationId: readPeerPublishedCollections
      security: [{nodeCredential: []}]
      x-ddp-node-credential-operation: directory_collections_read
      description: |
        Node-entry peer read served by control-api. Control verifies a single-use signed
        node credential and its query digest, proxies to its own corpus with local service identity,
        and re-checks that every descriptor is originated by this node before any peer sees
        it. Caller scope, caller digest and readiness internals are never exposed.
      parameters:
        - {name: snapshot_id, in: query, schema: {type: string, maxLength: 32}}
        - {name: cursor, in: query, schema: {type: string, maxLength: 32}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
      responses:
        '200': {description: Stable node catalog page; continuation cannot change the page size, content: {application/json: {schema: {$ref: '#/components/schemas/PeerCatalogPage'}}}}
        '401': {description: 'credential_invalid, credential_expired or credential_replayed'}
        '403': {description: 'node_unknown, node_revoked, credential_audience_mismatch or credential_scope_denied'}
        '409': {description: wrong_target or snapshot revision conflict}
        '410': {description: Peer catalog snapshot expired or withdrawn; already observed targets are retained by the caller}
        '502': {description: Upstream corpus producer unavailable or inconsistent}
        '503': {description: Node identity or corpus connection unavailable}
        '507': {description: Authorized catalog too large; never claim complete}
components:
  securitySchemes:
    session: {type: http, scheme: bearer}
    userKey: {type: apiKey, in: header, name: X-API-Key}
    service: {type: http, scheme: bearer}
    nodeCredential: {type: apiKey, in: header, name: X-DDP-Node-Credential, description: Single-use Ed25519 directory credential with live approval and durable replay protection}
...
  schemas:
...
    CatalogPage:
      type: object
      properties:
        snapshot_id: {type: string, pattern: '^[0-9a-f]{32}$'}
        scope_id: {type: string, minLength: 1, maxLength: 128}
        caller_scope_hash: {type: string, pattern: '^sha256:[0-9a-f]{64}$'}
        origin_node_id: {$ref: '../schemas/ddp-discovery/v1.json#/$defs/node_id'}
        registry_revision: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        valid_until: {type: string, format: date-time}
        first_cursor: {type: string, pattern: '^[0-9a-f]{32}$'}
        terminal_cursor: {type: string, pattern: '^[0-9a-f]{32}$'}
        total: {type: integer, minimum: 0, maximum: 10000}
        collections: {type: array, maxItems: 100, items: {$ref: '../schemas/ddp-discovery/v1.json#/$defs/CollectionDescriptor'}}
        index_readiness: {type: object, additionalProperties: {type: string, enum: [ready, unavailable]}}
        next_cursor: {type: [string, 'null'], pattern: '^[0-9a-f]{32}$'}
        complete: {type: boolean}
        content_snapshot_complete: {const: false}
      required: [snapshot_id, scope_id, caller_scope_hash, origin_node_id, registry_revision, created_at, valid_until, first_cursor, terminal_cursor, total, collections, index_readiness, next_cursor, complete, content_snapshot_complete]
      additionalProperties: false
      allOf:
        - if: {properties: {complete: {const: true}}}
          then:
            properties: {collections: {maxItems: 0}, next_cursor: {const: null}, index_readiness: {maxProperties: 0}}
    PeerCatalogPage:
      type: object
      properties:
        authority_node_id: {$ref: '../schemas/ddp-discovery/v1.json#/$defs/node_id'}
        snapshot_id: {type: string, pattern: '^[0-9a-f]{32}$'}
        registry_revision: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        valid_until: {type: string, format: date-time}
        first_cursor: {type: string, pattern: '^[0-9a-f]{32}$'}
        terminal_cursor: {type: string, pattern: '^[0-9a-f]{32}$'}
        total: {type: integer, minimum: 0, maximum: 10000}
        collections: {type: array, maxItems: 100, items: {$ref: '../schemas/ddp-discovery/v1.json#/$defs/CollectionDescriptor'}}
        next_cursor: {type: [string, 'null'], pattern: '^[0-9a-f]{32}$'}
        complete: {type: boolean}
      required: [authority_node_id, snapshot_id, registry_revision, created_at, valid_until, first_cursor, terminal_cursor, total, collections, next_cursor, complete]
      additionalProperties: false
      allOf:
        - if: {properties: {complete: {const: true}}}
          then:
            properties: {collections: {maxItems: 0}, next_cursor: {const: null}}